refactor(platform): move the event vocabulary out from behind the window feature
Naming a key is not the same as opening a window. WindowEvent, KeyCode,
PointerButton and the event shape table were behind the window feature
alongside the code that drives the OS loop, so any consumer that merely
speaks the vocabulary had to compile a windowing library to name a key.

They now live in renew_platform::event, which is always available; the
window module re-exports them, so existing paths resolve unchanged. Only
the loop-driving half stays gated.

The input layer is the crate this was costing: it takes the platform
with default features off now and pulls no window system at all. A build
cell guards that absence rather than trusting it.

The shape-table guard moves with the types it guards, so the headless
configuration runs it too -- it did not before.
